+++ revival/init.lua
@@ -1,4 +1,4 @@
-local x = os.clock()
+--local x = os.clock()
 
 revival = {}
 craft = {}
@@ -12,5 +12,7 @@ dofile(mod_path .. "/items.lua")
 dofile(mod_path .. "/tools.lua")
 dofile(mod_path .. "/crafting.lua")
 dofile(mod_path .. "/dissection_table.lua")
+dofile(mod_path .. "/manual_press.lua")
+dofile(mod_path .. "/override.lua")
 
-minetest.log("action", string.format("elapsed time: %.2f\n", os.clock() - x))
+--minetest.log("action", string.format("elapsed time: %.2f\n", os.clock() - x))