+++ revival/override.lua
@@ -1,0 +1,6 @@
+minetest.override_item(
+    "default:apple",
+    {
+        groups = {pressable = 1}
+    }
+)